Initialize Sopel DNS plugin, based on code pasted by catbeard
On January 13, 2021, a user named catbeard sent a link to a pastebin
containing code for this plugin to #sopel at freenode:

  [15:04:56] <catbeard> this one's pretty handy so far
  [15:04:59] <catbeard> https://paste.sh/sERHybXg#LY5MeTWW4OJCXDVSSK3Ofaw5

I suggested releasing this on PyPI and/or having the sopel-irc GitHub
organization adopt such a plugin, but never got a reply.

Since I want to have such a tool in my own bot, here's a package. Some
tweaks have been made to the original code pasted, and I plan to add
more record types in the future as a fun little side project.